+++ index.js
@@ -37,10 +37,12 @@ if (cluster.isMaster) {
     const client = new Discord.Client();
     const updateJsonFile = require('update-json-file');
     const cards = require('./persistant_data/MonopolyCards.json');
-    const roleFunction = require('./modules.js');
+    const roleFunction = require('./modules/roles.js');
+    const game = require('./modules/game.js');
+    const playerData = require('./modules/playerData.js');
 
     let fs = require('fs');
-    let game = false;
+    let gameStart = false;
     let playercount = 0;
     let joinable = true;
     let lobby = [],
@@ -50,6 +52,7 @@ if (cluster.isMaster) {
 
     /*Start bot */
     client.login(auth.token);
+
     client.on("ready", () => {
         console.log(`Bot has started, with ${client.users.size} users, in ${client.channels.size} channels of ${client.guilds.size} guilds.`);
         const member = client.channels.get(config.statusChannelID);
@@ -67,9 +70,12 @@ if (cluster.isMaster) {
 
     /* Will run when it sees a message */
     client.on("message", async message => {
+      // Will read the table for each role tag and put it in lobbies[]; in ./modules/playerdata.js
+
+
         /* Will ignore it self */
 
-        //if (message.author.bot) return;
+        if (message.author.bot) return;
 
         //Will search for the prefix for the bot to function
         if (message.content.indexOf(config.prefix) !== 0) return;
@@ -88,17 +94,12 @@ if (cluster.isMaster) {
             case 'r':
                 console.log("Reload time!");
                 process.exit(0);
-                break;
-            case 'ping':
-              return message.channel.send("pong");
             break;
-
-            case 'testcards':
-                var code = "message.channel.send(cards.";
-                code = code + combWord + ")";
-                 eval(code);
-            break;
-
+            // for testing
+            case 'g':
+              playerData.setPlayerInJail(message,CGID[CGID.length-1]);
+              //playerData.setBalance(message, CGID[CGID.length - 1]);
+            break;was
             case 'roll':
               let roll1 = Math.floor(Math.random()*7);
               let roll2 = Math.floor(Math.random()*7);
@@ -115,28 +116,29 @@ if (cluster.isMaster) {
 
             case 's':
             case 'start':
-              if (game)
+              if (gameStart)
                 return message.channel.send("game is already started please don't start another game.");
-              game = true;
+              gameStart = true;
               message.channel.send ("Game has started.")
-            
+
               // Purpose of the loop is to check if the randomly generated CGID has been used yet, if so make a new ID
-              CGID[CGID.length - 1] = roleFunction.getNewCGID(CGID);
-              
+              CGID[CGID.length] = roleFunction.getNewCGID(CGID);
+              let f = CGID[CGID.length-1].toString()
+              console.log(CGID);
               // Will create a role with the last number generated.
               message.guild.createRole({
                 name: `lobby#${CGID[CGID.length-1]}`,
                 hoist: false,
                 mentionable: false,
               })
-                .then( () => roleFunction.createChannel(message, `lobby#${CGID[CGID.length - 1]}`));
-              
+                .then( () => roleFunction.createChannel(message, `lobby#${CGID[CGID.length - 1]}`))
+
             break;
 
             case 'join':
             case 'play':
               // Checks to see if the game has to started if not then no one can join.
-              if(!game)
+              if(!gameStart)
                 return message.channel.send ("Please start the game before joining.");
               // current max lobby limit is 10, should let the server owner change it in the future.
               if ( lobby.length < 10){
@@ -153,8 +155,9 @@ if (cluster.isMaster) {
                 let role = message.guild.roles.find(r => r.id !== message.guild.id && r.name == roleName);
 
                 guildMember.addRole(role).catch(()=>console.error("adding role"));
+                playerData.setPlayerData(message, CGID[CGID.length-1].toString());
               }
-            
+
               if (lobby.length > 0) {
               // If the player count is greater 1 then the game will start in x amount of minutes.
                 setTimeout(function(){
@@ -164,14 +167,19 @@ if (cluster.isMaster) {
                   }
                   players = players.substring(0, players.length - 1);
                   message.reply("Game has started! Enjoy!!");
-                  
-                  playGame(players)
-                  
-                  game = false;
+
+                  // playGame(players)
+
+                  gameStart = false;
                   lobby = [];
                 10}, 6000)
                 }
 
+            break;
+            // only here for testing.
+            case 'makeplayer':
+                game.createPlayer( message, CGID[CGID.length - 1], client );
+                game.testDb();
             break;
 
             case 'endgame':
@@ -188,8 +196,8 @@ if (cluster.isMaster) {
               }
 
               if (roleName === undefined)
-                return message.channel.send("Sorry that game number does not exist.");      
-              
+                return message.channel.send("Sorry that game number does not exist.");
+
               let role = message.guild.roles.find(r => r.name == roleName);
               // Deletes the roll after it has been found.
                 role.delete('good night')
@@ -197,35 +205,39 @@ if (cluster.isMaster) {
                   .catch(console.error);
               roleFunction.deleteChannel(message, args[0], false)
 
+              gameStart = false;
             break;
 
             default:
                 return message.reply("That is no command.");
             break;
-                
+
             case'endall':
               let lobbyRole = roleFunction.getLobbyRole(message);
               console.log("I got lobbyRole");
               let rolep = '';
-               for(let i = 0 ; i < lobbyRole.length; i ++){
-                rolep = message.guild.roles.find(r => r.name == lobbyRole[i]);
-                rolep.delete('good night')
-                  .then(deleted => console.log(`Deleted role ${deleted.name}`))
-                  .catch(console.error);
-              }
+              roleFunction.deleteChannel(message, -1, true)
+                for (let i = 0; i < lobbyRole.length; i++) {
+                  rolep = message.guild.roles.find(r => r.name == lobbyRole[i]);
+                  rolep.delete('good night')
+                    .then(deleted => console.log(`Deleted role ${deleted.name}`))
+                    .catch(console.error);
+                }
+
+
               message.channel.send ("All lobby roles have been deleted.");
-              game = false;
-              roleFunction.deleteChannel(message,-1,true)
+              gameStart = false;
+
             break;
         }
-        
+
     });
-      async function playGame(UUID_String) {
-      let players = []
-      players = UUID_String.split(",");
-      for (let i = 0 ; i < players.length; i ++) {
-        let debug = client.channels.get(411777884861104130)
-        debug.send(players[i])
-      }
-    }
+    //   async function playGame(UUID_String) {
+    //   let players = []
+    //   players = UUID_String.split(",");
+    //   for (let i = 0 ; i < players.length; i ++) {
+    //     let debug = client.channels.get(411777884861104130)
+    //     debug.send(players[i])
+    //   }
+    // }
 }